% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseScalars.R
\name{baseScalars}
\alias{baseScalars}
\title{baseScalars}
\usage{
baseScalars(baseyr = 2020)
}
\arguments{
\item{baseyr}{Base year for IMPACT simulation. Currently 2020}
}
\value{
Hunger scalars for pop at risk calculation
}
\description{
baseScalars
}
\examples{
\dontrun{
baseScalars()
}
}
