% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAggregate.R
\name{calcAggregate}
\alias{calcAggregate}
\title{calcAggregate}
\usage{
calcAggregate(gdx, baseyr, mapping = c("aggreg_standard.csv"))
}
\arguments{
\item{gdx}{GDX file from an IMPACT run}

\item{baseyr}{Base year for simulation}

\item{mapping}{Vector of mapping files with their extension. Defaults to
"aggreg_standard.csv". You can add multiple mappings with}
}
\value{
Aggregate indicators for Hunger and Nutirition
}
\description{
calcAggregate
}
\examples{
\dontrun{
calcAggregate()
}
}
