% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoreCalc.R
\name{CoreCalc}
\alias{CoreCalc}
\title{Core Calculations for this module}
\usage{
CoreCalc(gdx, baseyr = 2020)
}
\arguments{
\item{gdx}{GDX file from an IMPACT run}

\item{baseyr}{Base year for simulation. Currently 2020.}
}
\value{
Core calculation return
}
\description{
Core Calculations for this module
}
\examples{
\dontrun{
CoreCalc(gdx)
}
}
