% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDrivers.R
\name{readDrivers}
\alias{readDrivers}
\title{readDrivers}
\usage{
readDrivers(gdx, var)
}
\arguments{
\item{gdx}{GDX file from an IMPACT run}

\item{var}{Which variable to read. Use "pop" or "gdp" or "food"}
}
\value{
Returns population, GDP(pc) or food availability
}
\description{
readDrivers
}
\examples{
\dontrun{
readDrivers(gdx, var = "pop")
}
}
